Create lilibyte.yaml
"Everything is in the credits. Now that society has been definitively turned into an enterprise, everything is in the synopsis of performance and enterprise, and its leaders must produce all the signs of the advertising ‘look’. The slightest failing becomes unpardonable, since the whole nation would be diminished by it. ... By contrast, political weaknesses or stupidity are of no importance. Image alone counts."