Cancel the Cancel Culture
I was on the edge about signing simply because there is not much obvious evidence for or against him at this time.  The RMS ~~Open~~ Cancel Letter is intentionally vague about the details and would much rather call him all the trigger words they know to paint someone as the boogie-man.

Let's not let them cancel.

Also the cancel letter is on github.com/trending, but the support letter isn't despite similar activity...

Also, who *actually* thinks they can somehow get the entire board of an organization to resign over harmless debate ramblings of one individual.  That is the stupidest thing I have ever heard of.