Dumitru Uzun - duzun
Who is next? Torvalds?
The truth is uncovered to the ones that pursuit truth by any means and has the courage to accept it.
RMS is a great man and a pylon of the Open Source community and philosophy! 
Be careful, this is not an attack on one person, this is an ideological attack on the OS community and principle!