Create 1-7-1.yaml
RMS is a pioneer of the free software movement and philosophy as well as a legendary programmer. He is a historical figure and his legacy speaks for itself. I don't agree with everything he's ever said, but when it comes to free software, he's always been on the right track and his words are still spot on to this very day. To try and get rid of Richard Stallman is a perverse act of historic revisionism and an thinly veiled attack on an icon of liberty.