Create JonathinR.yaml
As a proponent of FOSS, I can not in good conscience allow a prominent figure and founder of the Free Software Foundation to be removed. If he is to be removed, this will throw away decades of forward momentum for the Free Software Movement and slow the widespread adoption of free and open source software. If someone disagrees with his opinions on something, they should not seek to destroy something he's a part of just for the sake of tearing him down.