Create jschwartzenberg.yaml
I went through the letters. The unfortunate letter against RMS writes "RMS recommends" and clicking the link behind it there is written "I think". I see a pattern of twisted words to try make a statement, but this is not an appropriate way to misrepresent someone even when that person has made make big mistakes otherwise. Addressing issues becomes harder when the circumstances are dishonest. The unfortunate letter creates an atmosphere in which people who have valid criticism or who have been mistreated, face extra difficulty to stand out and have their concerns addressed. I encourage independent concerns to be addressed as such and to avoid any association with the unfortunate letter which I'm fearing is already mistakenly being done by some supporters of this letter.