Create ForgottenUmbrella.yaml
Having differing opinions, completely unrelated to your position at an organisation, should not be a valid reason for people to throw a fit calling for your resignation. Judge someone on their skill at the job, not their views on life.
On a slightly unrelated note, I like pretending to have privacy and anonymity, hence the use of a pseudonym. I hope that is okay.